refactor(client): reduce imports in list/index route pages
The six list/index routes each tripped import/max-dependencies because
they re-composed the same page chrome (header, states, filter bar) and a
per-entity box. Introduce a listControls barrel (chrome) and a boxes
barrel so a page pulls these from one module instead of five, and lift
the heavy filter/sort bodies of the three worst offenders into
presentational route-private *-List components (data and callbacks via
props; queries/mutations stay in the route). Presentational extraction
also lets the new components be exercised in Storybook with the existing
props + RouterStub convention.

All six routes and the three new components now sit at <=10 value
imports; typecheck, lint, and the client test suite (incl. the new story
play tests) stay green.
