+++ packages/client/src/components/listControls/index.ts
@@ -1,0 +1,19 @@
+// Single import surface for the chrome shared by list/index route pages
+// (page header, route loading/error/empty states, view toggle, and the filter
+// bar primitives). Aggregates components that live elsewhere — keep the
+// implementations in their original homes; this only re-exports them so a list
+// page pulls them from one module instead of five.
+export { EntityError, EntityPending } from "@/components/EntityStates";
+export { OnboardingEmptyState } from "@/components/layout/OnboardingEmptyState";
+export { PageHeader } from "@/components/layout/PageHeader";
+export {
+  ViewModeToggle,
+  type ViewMode,
+} from "@/components/layout/ViewModeToggle";
+export {
+  ClearFiltersButton,
+  FilterSelect,
+  type FilterSelectOption,
+  ListEmptyStates,
+  ListSearchInput,
+} from "@/components/ListPageControls";
